More splitting of test_atomgroup
