+++ testsuite/MDAnalysisTests/test_atomgroup.py
@@ -166,107 +166,6 @@ class TestUniverseSetTopology(object):
         assert_equal('improperDict' in self.u._cache, False)
 
 
-class TestSegmentGroup(object):
-    # VALID
-    @dec.skipif(parser_not_found('DCD'),
-                'DCD parser not available. Are you using python 3?')
-    def setUp(self):
-        """Set up the standard AdK system in implicit solvent."""
-        self.universe = MDAnalysis.Universe(PSF, DCD)
-        self.g = self.universe.atoms.segments
-
-    # VALID
-    def test_newSegmentGroup(self):
-        """test that slicing a SegmentGroup returns a new SegmentGroup (Issue 135)"""
-        g = self.universe.atoms.segments
-        newg = g[:]
-        assert_equal(type(newg), type(g), "Failed to make a new SegmentGroup: type mismatch")
-        assert_equal(len(newg), len(g))
-
-    # VALID
-    def test_n_atoms(self):
-        assert_equal(self.g.n_atoms, 3341)
-
-    # VALID
-    def test_n_residues(self):
-        assert_equal(self.g.n_residues, 214)
-
-    # INVALID: `SegmentGroup.resids` gives list of arrays, one array for each segment
-    @skip
-    def test_resids_dim(self):
-        assert_equal(len(self.g.resids), len(self.g.residues))
-
-    # INVALID: topology has no resnums
-    @skip
-    def test_resnums_dim(self):
-        assert_equal(len(self.g.resnums), len(self.g.residues))
-
-    # VALID
-    def test_segids_dim(self):
-        assert_equal(len(self.g.segids), len(self.g))
-
-    # INVALID: cannot set resids from `SegmentGroup`; no `set_resids` method
-    @skip
-    def test_set_resids(self):
-        g = self.universe.select_atoms("bynum 12:42").segments
-        resid = 999
-        g.set_resids(resid)
-        # check individual atoms
-        assert_equal([a.resid for a in g.atoms],
-                     resid * np.ones(g.n_atoms),
-                     err_msg="failed to set_resid for segment to same resid")
-        # check residues
-        assert_equal(g.residues.resids, resid * np.ones(g.n_residues),
-                     err_msg="failed to set_resid of segments belonging to atoms 12:42 to same resid")
-
-    # INVALID: cannot set resids from `SegmentGroup`; no `set_resids` method
-    @skip
-    def test_set_resids(self):
-        g = self.universe.select_atoms("resid 10:18").segments
-        resid = 999
-        g.set_resids(resid * np.ones(len(g)))
-        # note: all is now one residue... not meaningful but it is the correct behaviour
-        assert_equal(g.resids, [resid],
-                     err_msg="failed to set_resid  in Segment {0}".format(g))
-
-    # INVALID: no `set_segids` method; use `segids` property directly
-    @skip
-    def test_set_segids(self):
-        s = self.universe.select_atoms('all').segments
-        s.set_segids(['ADK'])
-        assert_equal(self.universe.segments.segids, ['ADK'],
-                     err_msg="failed to set_segid on segments")
-
-    # INVALID: no `set_segids` method; use `segids` property directly
-    @skip
-    def test_set_segid_updates_self(self):
-        g = self.universe.select_atoms("resid 10:18").segments
-        g.set_segids('ADK')
-        assert_equal(g.segids, ['ADK'],
-                     err_msg="old selection was not changed in place after set_segid")
-
-    # INVALID: no `set_segids` method; use `segids` property directly
-    @skip
-    def test_set_masses(self):
-        g = self.universe.select_atoms("bynum 12:42 and name H*").segments
-        mass = 2.0
-        g.set_masses(mass)
-        # check individual atoms
-        assert_equal([a.mass for a in g.atoms],
-                     mass * np.ones(g.n_atoms),
-                     err_msg="failed to set_mass in segment of  H* atoms in resid 12:42 to {0}".format(mass))
-
-    # INVALID: no `set_segids` method; use `segids` property directly
-    @skip
-    def test_set_segid_ValueError(self):
-        assert_raises(ValueError, self.g.set_resids, [1, 2, 3, 4])
-
-    # VALID
-    def test_atom_order(self):
-        assert_equal(self.universe.segments.atoms.indices,
-                     sorted(self.universe.segments.atoms.indices))
-
-
 class _WriteAtoms(object):
     """Set up the standard AdK system in implicit solvent."""
     ext = None  # override to test various output writers
